+++ src/entity/episode/episode.args.ts
@@ -1,6 +1,6 @@
 import {ObjField, ObjParamsType} from '../../modules/rest/decorators';
 import {EpisodeOrderFields, ListType, PodcastStatus} from '../../types/enums';
-import {ArgsType, Field, ID, InputType, Int} from 'type-graphql';
+import {ArgsType, Field, Float, ID, InputType} from 'type-graphql';
 import {OrderByArgs, PaginatedArgs} from '../base/base.args';
 import {examples} from '../../modules/engine/rest/example.consts';
 
@@ -41,7 +41,7 @@ export class EpisodeFilterArgs {
 	@ObjField(() => [String], {nullable: true, description: 'filter by Podcast Ids', isID: true})
 	podcastIDs?: Array<string>;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	@ObjField({nullable: true, description: 'filter by Creation timestamp', min: 0, example: examples.timestamp})
 	since?: number;
