+++ src/entity/podcast/podcast.args.ts
@@ -1,6 +1,6 @@
 import {ObjField, ObjParamsType} from '../../modules/rest/decorators';
 import {examples} from '../../modules/engine/rest/example.consts';
-import {ArgsType, Field, ID, InputType, Int} from 'type-graphql';
+import {ArgsType, Field, Float, ID, InputType} from 'type-graphql';
 import {ListType, PodcastOrderFields, PodcastStatus} from '../../types/enums';
 import {FilterArgs, OrderByArgs, PaginatedArgs} from '../base/base.args';
 
@@ -51,7 +51,7 @@ export class PodcastFilterArgs {
 	@ObjField(() => [String], {nullable: true, description: 'filter by Episode Ids', isID: true})
 	episodeIDs?: Array<string>;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	@ObjField({nullable: true, description: 'filter by Creation timestamp', min: 0, example: examples.timestamp})
 	since?: number;
 
@@ -63,11 +63,11 @@ export class PodcastFilterArgs {
 	@ObjField(() => [PodcastStatus], {nullable: true, description: 'filter by Podcast Status', example: [PodcastStatus.downloading]})
 	statuses?: Array<PodcastStatus>;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	@ObjField({nullable: true, description: 'filter by since Last Check timestamp', min: 0, example: examples.timestamp})
 	lastCheckFrom?: number;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	@ObjField({nullable: true, description: 'filter by until Last Check timestamp', min: 0, example: examples.timestamp})
 	lastCheckTo?: number;
 
@@ -106,7 +106,7 @@ export class PodcastFilterArgsQL {
 	@Field(() => [ID], {nullable: true})
 	episodeIDs?: Array<string>;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	since?: number;
 
 	@Field(() => String, {nullable: true})
@@ -115,10 +115,10 @@ export class PodcastFilterArgsQL {
 	@Field(() => [PodcastStatus], {nullable: true})
 	statuses?: Array<PodcastStatus>;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	lastCheckFrom?: number;
 
-	@Field(() => Int, {nullable: true})
+	@Field(() => Float, {nullable: true})
 	lastCheckTo?: number;
 
 	@Field(() => String, {nullable: true})
